WMV-363 : fix last compilation warnings after moving to x64

+++ src/core/GameDatabase.h
@@ -60,7 +60,7 @@ namespace core
     QString type;
     bool isKey;
     bool needIndex;
-    size_t arraySize;
+    uint arraySize;
     int id;
   };
